+++ docsig/_decorators.py
@@ -10,16 +10,17 @@ from pathlib import Path as _Path
 from .messages import E as _E
 
 _FuncType = _t.Callable[..., _t.Union[int]]
+_WrappedFuncType = _t.Callable[..., _t.Union[str, int]]
 
 
-def parse_msgs(func: _FuncType) -> _FuncType:
+def parse_msgs(func: _WrappedFuncType) -> _WrappedFuncType:
     """Parse error codes or symbolic messages into message objects.
 
     :param func: Function to wrap.
     :return: Wrapped function.
     """
 
-    def _wrapper(*args: str | _Path, **kwargs: _t.Any) -> int:
+    def _wrapper(*args: str | _Path, **kwargs: _t.Any) -> str | int:
         disable = list(_E.fromcodes(kwargs.get("disable", []))) or None
         targets = list(_E.fromcodes(kwargs.get("targets", []))) or None
         kwargs["disable"] = disable
@@ -29,31 +30,27 @@ def parse_msgs(func: _FuncType) -> _FuncType:
     return _wrapper
 
 
-def validate_args(func: _FuncType) -> _FuncType:
+def validate_args(func: _FuncType) -> _WrappedFuncType:
     """Confirm args passed to function are valid.
 
     :param func: Function to wrap.
     :return: Wrapped function.
     """
 
-    def _wrapper(*args: str | _Path, **kwargs: _t.Any) -> int:
+    def _wrapper(*args: str | _Path, **kwargs: _t.Any) -> str | int:
         if not kwargs.get("list_checks", False):
-            if not args and not kwargs.get("string", None):
-                raise SystemExit(
-                    "the following arguments are required: path(s) or string",
+            if not args and not kwargs.get("string"):
+                return (
+                    "the following arguments are required: path(s) or string"
                 )
 
             for message in kwargs.get("disable") or []:
                 if not message.isknown:
-                    raise SystemExit(
-                        f"unknown option to disable '{message.description}'",
-                    )
+                    return f"unknown option to disable '{message.description}'"
 
             for message in kwargs.get("targets") or []:
                 if not message.isknown:
-                    raise SystemExit(
-                        f"unknown option to target '{message.description}'",
-                    )
+                    return f"unknown option to target '{message.description}'"
 
         return func(*args, **kwargs)