+++ tests/conftest.py
@@ -2,6 +2,8 @@
 tests.conftest
 ==============
 """
+from __future__ import annotations
+
 from pathlib import Path
 
 import pytest
@@ -32,7 +34,7 @@ def fixture_main(monkeypatch: pytest.MonkeyPatch) -> MockMainType:
     :return: Function for using this fixture.
     """
 
-    def _main(*args: str) -> int:
+    def _main(*args: str) -> str | int:
         """Run main with custom args."""
         monkeypatch.setattr(
             "sys.argv",